Use the right class for the message to be sent
Starting from Moodle 2.9, the stdClass should not be used for the
messages. The support for using it was dropped in Moodle 3.6.

In newer versions, no message was sent to the admin and instead the
following error is thrown:

```
availability_paypal IPN exception handler: Exception - Argument 1 passed
to message_send() must be an instance of core\message\message, instance
of stdClass given, called in /availability/condition/paypal/ipn.php on line 217
```

References:

* https://tracker.moodle.org/browse/MDL-55449
* https://docs.moodle.org/dev/Message_API#How_to_send_a_message
* https://docs.moodle.org/dev/Moodle_2.9_release_notes#For_developers