Fix missing messaging providers and introduce new capability
The messaging providers were not declared in the db/messages.php. As a
result, attempting to call message_send() only let to error in the log:

"Attempt to send msg from a provider availability_paypal/payment_error
that is inactive or not allowed for the user id=xxx"

and no actual message / emails was sent.

This commit declares the messaging providers and also introduces a new
capability that will be used to control who receives payment errors.