Improve the landing page after returning from PayPal
The user can easily return back to Moodle before the IPN is processed.
In that case we still to make it look as a pending payment, even if we
still do not have the "Pending" transaction created in the database.

The patch introduces a token "payment identifier" that allows us to
detect that they were just redirected back from PayPal.

Also a new "Continue" button is added to the landing page so users have
easier way to navigate away from the page.