Better identification of the incoming IPN
If the same PayPal account is used for multiple purposes, it always
sends the IPN to our listener - even if the payment was not initiated by
the plugin.

This leads to false emails and has potential to break things in
unexpected way.

The patch adds the plugin's identifier to the custom IPN parameter so
that the listener only processes the transactions originated in it.